requirements.txt
alembic==1.9.2
attrs==22.2.0
catboost==1.0.4
category-encoders==2.6.0
cloudpickle==2.2.1
cmaes==0.9.1
colorlog==6.7.0
colour==0.1.5
contourpy==1.0.7
cycler==0.11.0
DateTime==5.0
dtreeviz==2.1.3
exceptiongroup==1.1.0
fonttools==4.38.0
graphviz==0.20.1
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
joblib==1.2.0
kiwisolver==1.4.4
lightgbm==3.3.5
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
mljar-supervised==0.11.5
numba==0.56.4
numpy==1.24.1
optuna==3.1.0
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
plotly==5.13.0
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
scikit-learn==1.2.1
scikit-plot==0.3.7
scipy==1.10.0
seaborn==0.12.2
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
SQLAlchemy==2.0.0
statsmodels==0.13.5
tabulate==0.9.0
tenacity==8.1.0
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
typing-extensions==4.4.0
wordcloud==1.8.2.2
xgboost==1.7.3
zipp==3.12.0
zope.interface==5.5.2

